Take advantage of retina screen by iterating over pixels

+++ RpnCalc/AxesDrawer/AxesDrawer.h
@@ -14,11 +14,11 @@
          originAtPoint:(CGPoint)axisOrigin
                  scale:(CGFloat)pointsPerUnit;
 
-+ (float)convertToAxesCoordinates:(int) viewValue
++ (float)convertToAxesCoordinates:(float) viewValue
                            atOrigin:(float)origin
                           withScale:(float)scale;
 
-+ (int)convertToViewCoordinates:(float)axesValue
++ (float)convertToViewCoordinates:(float)axesValue
                         atOrigin:(float)origin
                        withScale:(float)scale;
